Create a Vale test for ease-of-use words
Words describing ease-of-use shouldn't appear in documentation.
While that statement may seem outrageous and odd from a Product
view, it's an empathy issue at heart. As technical writers, we're
aware that what's easy for an experienced user can be complex and
frustrating (dare I say "hard"?) for new users, and telling a user
a seemingly-impossible task is "easy" causes cognitive dissonance
and frustration.