Add repositionImageDiffNote mutation
This mutation allows repositioning a DiffNote.

It uses the special `reposition_note` permission, which is an alias of
`admin_note` unless the noteable is a Design, in which case it will be
true if the user can `create_note`.

This allows people who can comment on a Design to reposition the notes
of other people.

https://gitlab.com/gitlab-org/gitlab/-/issues/207334