Remove unused function in import mutations
It looks like usage of this function was removed in
a6e3e394952bd3e712d2889077aeb58c81f7d0fc but this function was missed in
the cleanup. Remove it so that it's more obvious what's acutally being
used in this file.

MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/45783