Fix issue with case sensitive search for projects
Prior to this change, if a user tried to search for projects to import,
they would have to match the case of the project name.

This changes the filter to use a case insensitive comparison so that the
user does not need to get the case right.

MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/45783
Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/36556