Improve clarity of the Rate Limits UI
Previously the admin interface for "User and IP Rate Limits" had inputs
with duplicate labels. These inputs were only differentiated by their
groupings on the page, which themselves were not clear. This improves
the input labels, makes the checkbox labels bold (as they are the most
important inputs in this section), and adds some horizontal rules to
visually distinguish the groupings. These changes are intended to make
this UI easier to understand, and match up better with the docs.