Verify that rack attack logging does not perform additional queries
There is a user lookup in the rack attack logging code, but if there is
a user, we will have already looked up their details earlier in the
request cycle. That means that this will use the ActiveRecord cache and
not actually hit the database again.

This spec ensures we don't add more queries when logging than we do when
simply running normally.