Install libpcre2-dev via package manager
Our installation documentation instructs users to install libpcre2 by
manually compiling it. This has been required in the past as it wasn't
available in some conservative distributions like Debian, but nowadays
it is available. Most importantly, Ubuntu 16.04 and Debian 9 both have
it in their repositories, which are the oldest versions which are still
supported.

Drop these instructions in favor of installing libpcre2-dev via the
package manager.