Avoid N+1 queries in JiraConnect::client
We expose a user notes count for each MR. Counting
user notes executes an extra SQL query for each MR in the list.
This commit loads all counts in one query and passes it to the
serializer as an option. If the option is not available, it will
fall back to execute a count query.