Global Search - Radio Filters
This change is behind a feature
flag :search_facets.

This was split off for the sake
of MVC and the final product
can be seen here:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/46595

This change replaces the dropdown
filter with radio buttons.

These will then be used as
the components in the
sidebar in the next MR.