Remove duplicate period in GitLab emails
A line can appear at the bottom of GitLab notification emails:
"Reply to this email directly or view it on GitLab". Previously that
line ended with two periods, because there is a period in the translation
as well as in the HAML template. This removes the period from the HAML
template so now there is only one period in the rendered email.