Move SAST Vuex module to CE
This moves (most of) the SAST Vuex module from EE to CE as part of the
effort to [add vulnerability counts to the CE MR security
widget][issue].

It's not a straight-forward move, though. The SAST module contains
actions, mutations and getters specific to EE functionality, so the
approach taken here was to move all _common_ behaviour to CE version,
and make the EE version _extend_ the CE version.

[issue]: https://gitlab.com/gitlab-org/gitlab/-/issues/273423

Co-authored-by: Jannik Lehmann <jlehmann@gitlab.com>