Update Terraform/Base.latest.gitlab-ci.yml to inline cd TF_ROOT
At present this template appears to be designed to never interfere with other
jobs in your CI YML. This is true for all cases except the global
`before_script` which will cause other jobs in your `.gitlab-ci.yml` to
possibly end up in the wrong directory. This was specifically important for a
use case where a project was using terraform and ansible in
https://gitlab.com/gitlab-org/quality/gitlab-environment-toolkit . The simplest
solution appears to be to just inline the `cd ${TF_ROOT}`. We also considered
defining another hidden job that is extended by all the other jobs here but the
indirection didn't seem worth it for a single step.