Replace deprecated modal in detailed_metric
The component detailed_metric vue had a deprecated button
and modal which are now being replaced. We also changed
the way we toggle logs to get away from javascript
custom functions and instead rely on having the data
in Vue. Tests were added to ensure that everything works
as expected.