Fail verification timeouts
Verification concurrency is managed by the verification_started_at time.
If a `VerificationBatchWorker` or `VerificationWorker` is lost or
killed, then one or more records would be stuck in the
"verification_started" state. This worker unsticks those records. It is
called every minute.

Performance should be fine, but if desired, we could call this worker
directly with sidekiq-cron every hour instead of piggybacking on the
`trigger_background_verification` method, without noticeable impact.