Remove un-used endpoint for embeddable Insights
A few months ago there was interest in enhancing the Insights feature
to allow for these custom charts to be embedded in other contexts. As
a first step toward that, an endpoint was introduced for these
embeddable charts. However, it has been decided we have higher-priority
features to pursue. So this removes that endpoint code, however we
may choose to restore it in the future if we decide to pursue
embeddable Insights in the future.