Allow secondary emails in user search
User searching only happened with an equality check for the primary
email. We already had a search_with_secondary_emails helper however
inefficient because it was using an IN condition.

Since there is a unique constraint on secondary emails as well,
that can be updated with a strict equality check and make the query
perform similarly to the one without secondary emails check.

That allows us to default to searching through secondary emails as well
(when using the API, project member list etc).