Fix repository clone panel for wikis
The corresponding JS code in `pages/projects/project.js` got out of sync
and introduced a small bug where the dropdown label wasn't updated
correctly when switching the protocol.

Since this older version of the clone panel is only used for wikis, and
we also want to reuse it for group wikis outside of a project context,
the related code is moved into a wiki context, and the JS code is
copied/fixed in `pages/shared/wikis/wikis.js` to fix the mentioned bug.