Add started_at/completed_at timestamps to Elasticsearch migrations
When saving Elasticsearch migrations we set a started_at and
completed_at value. This will help us long term with debugging as this
information will make it much easier to find relevant logs for what
happened.

As this index was not declared with [strict dynamic
mapping](https://www.elastic.co/guide/en/elasticsearch/reference/7.x/dynamic.html)
it is totally safe to just add new fields to the payload without them
existing already in the configuration and I have tested locally that
this works as well.

For the sake of clarity, however, we have added them to the mappings as
there is no reason not to.