Mention plugins in troubleshooting Elasticsearch integration
This comes after a recent issue was diagnosed for one of our customers
where the plugin was causing the integration not to work. Hopefully
calling out plugins in our docs may save some time as customers may be
able to spot early on that the plugin is the thing causing issues before
asking us to diagnose the problem. Since there are many plugins that are
3rd party to Elasticsearch and they can significantly change the
behaviour of Elasticsearch it's impossible for us to test against all
plugins and be sure what does and does not work.