Validate relations before importing
Some project imports are missing Merge Requests because of some invalid
notes. It's preferable to lose a Merge Request note than a whole merge
request. Therefore, to avoid missing _major_ relations, every built
relation is being validated. This way, if a note is invalid, an
`ImportFailure` is logged, but the process continues for the
MergeRequest.