Improve some of our GraphQL helper interfaces
The pagination tests are updated to be clearer, and a missing pagination
test is added.

The sort_enum is fixed - previously the upper and lowercase forms were
not synonymous (one would have string values, and the other had symbol
values). Now both have symbol values, and a case where this caused a bug
(in the requirements_finder) was detected and fixed.

Since the pagination specs now expect actual values rather than strings
everywhere, a number of sort_param values have been updated to be
symbols, which map to graphql Enums.