Remove unused indexes
These indexes were originally added because it was assumed that
verification backfill queries would be the same for the new Geo
Self-Service Framework as they were for legacy project/wiki
verification. But we have added a state machine and are using a
different approach for the framework, so the queries are
different, and will not use these indexes.