Only run publish in npm.gitlab-ci.yml when needed
This commit updates the npm.gitlab-ci.yml template to only run the
`npm publish` command when it detects that the version in package.json
has not yet been published.

This commit also makes a number of other improvements to the template.
See https://gitlab.com/gitlab-org/gitlab/-/merge_requests/48615 for more
information.