Github Importer - import pull request reviews from Github
Github has the concept of pull request reviews, where a user can add a
comment with the type: "APPROVED", "CHANGES_REQUESTED" or "COMMENTED".

This commit introduces the ability to import this type of comments from
Github.

For the "APPROVED" type, if the author of the comment can be mapped to a
Gitlab user, an merge request approval is created.

For all the types a note is added to the merge request. If the author of
the review cannot be mapped to a Gitlab user a complementary text is
added to the original review body informing the original author of the
review.