Allow saving deprecated `import_sources`
Validation on `google_code` value in `import_sources` was enforced prior
to DB cleanup was executed. This causes issues with saving
`application_settings` on instance, including GitLab.com that still has
`google_code` as a value in `import_sources`.

This change is an interim step before a proper DB migration is
introduced to clean up the `import_sources`.