Github importer Avoid touching MR when importing merged by
When importing a Pull Request `merged by` field, if the user cannot be
mapped to a Gitlab user, a note is created, which `touch` the
MergeRequest (callback). This callback can be skipped by setting the
`importing` attribute of the note to `true`.

Related to: https://gitlab.com/gitlab-org/gitlab/-/issues/289671