Explicitly require public visibility of emails
The GitHub import performs a REST API call to fetch a user's email to
associate their content to their GitLab account.
The REST API of GitHub will show a `null` email address field if the
primary email address is not marked to be publicly visible in the
account's profile settings page.

Therefore the doc must (more) explicitly state the need for a publicly
visible email address on the GitHub end prior to the import.