Sort merge request diff files directory first
Each file browser and the diffs we show under the "Changes"
tab need to be consistent to prevent confusion.

Whenever we create a merge request diff, we create and store
`MergeRequestDiffFile` records and we keep the `relative_order`
of each file.

In order for the new sorting order to persist, we will sort the
diff files on:

- when a `MergeRequestDiff` is created
- when we load merge request diffs

Then we persist/update the `relative_order` of each file.