Speed up pull mirroring by getting updated tags straight from Gitaly
Currently, pull mirroring will fetch the full list of tags for a
repository twice - before and after the pull - and use the diff between
them to work out which new tags have been added.

Instead, we can parse the `git fetch` output to learn which tags have
been added, and use that to avoid pulling the list of tags twice.

We don't get information on the target of brand-new tags from git fetch
so we still need to pull the list of tags once, if new tags have been
created. However, this should still represent a significant improvement
for mirrored repositories with a large number of tags.