Add "Compliance Settings" section to documentation
GitLab frequently refers to "Compliance Settings" internally, but we
do not yet have a single source of truth (SSOT) that lists all of these
settings in one place. As GitLab's documentation is the expected location
for SSOT, this adds a section to the "Compliance Features" documentation
to list all "Compliance Settings". This is just a draft/concept and is
not yet complete.