Atomically select an available replica if all caught up
During a merge, we attempt to find a matching merge request using a
replica that should be up-to-date with the primary for a given WAL
location. However, there is a race condition that can happen if service
discovery alters the host list after this check has taken place.

To avoid this, we introduce an atomic selection process:

1. Check all the hosts that they have caught up to the given WAL location.
2. If all hosts have caught up, choose a host from that list.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/247857