Add support for fast forking for public projects
Previously when a public project was forked, it would do a full `git
clone` from the source repository. After that completed, then it would
link the parent's object pool as a Git alternate.

However, that is inefficient because the full repository still has to be
cloned. To speed this up, we can simply create an empty repository, link
the object pool as an object pool, and then fetch the references from
the source project. All objects will be present in the pool, and the
fork is fast because the objects do not have to be transmitted via a
full clone.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/24523