Disable epic quick actions for not persisted epics
Adding/removal of epic links expects that both parent and child epics
are created. If parent_epic is applied on a new not-persisted epic,
then EpicLink then this epic link is created before target epic is saved
which has couple of side-effects:
* CreateService's set_child_epic fails to update the new epic
* no system notes are created
* internal_id for epics is out-of-sync because EpicLink attempts to
  store new epic (and fails and rolls-back thsi save) -> then
  internal_id is already set which target epic is saved (as part of
  regular save operation) but not bumped in DB

Because epic quick actions don't work properly with unpersisted epics,
it will be disabled until the code is updated to work properly with new
epics.